Split ServerEmojiEvent out from ServerEvent class
